# compatibilities/open-api.spec.yaml
openapi: 3.0.1
info:
  title: Compatibiilties API
  description: A compatibilities API to serve the Angular front-end for matching like-minded people
  contact:
    name: Paul Costigan
    email: email@email.com
  license:
    name: ISC
    url: "https://opensource.org/licenses/ISC"
  version: 1.0.0
servers:
- url: http://localhost:3000/v1
  variables:
    environment: 
      default: localhost
      description: run the service in the local environment
    port:
      default: '3000'
      description: the port to be served in the environment
...
paths:
  /user:
    get:
      tags:
        - User
      summary: get all user information
      description: This endpoint will retrieve the user info
      operationId: getUser
      deprecated: false
      parameters:
      - in: path
        name: userId
        description: user id
        required: true
        allowEmptyValue: false
        schema:
          pattern: /^[a-f\d]{24}$/i
          type: string
      # - in: query
      #   name: interests
      #   schema:
      #     type: string
      #   description: "Select true to enhance the interests model in the response body"
       
      responses:
        200:
          description: Search executed successfully. If results were found they are
            returned as a list. If no results are found then an empty list is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseObject'
              example:
                message: "successfully retrieved user profile for: 5d30cc07ef01bfcd6a089338"
                response:
                  _id:  "5d30cc07ef01bfcd6a089338"
                  name: "John Smith"
                  age: 30
                  matches: []
                  interests: []
                  createdAt: "2019-07-18T19:44:07.672Z"
                  updatedAt: "2019-07-18T19:44:07.672Z"
                  __v: 0
        422:
          description: Invalide query parameter, incoreectly formatted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
              example:
                message: "invalid userId query parameter supplied: 5d30cc07ef01bfcd6a089338"
                error: {}
                  
        400:
          description: Query parameter incorrectly formatted or missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
              example:
                message: "there was an error with finding the user: 5d30cc07ef01bfcd6a089338, please try again"
                error: "the error stack"
        404:
          description: Query parameter incorrectly formatted or missing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseObject'
              example:
                message: "there is no profile for user: 5d30cc07ef01bfcd6a089338"
                response: {}
        500:
          description: Error while processing the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
              example:
                error: "there was a problem contacting the server, please try again"
    post:
      tags:
        - User
      summary: submits a new user
      description: This endpoint will retrieve the user info
      operationId: postUser
      deprecated: false
      requestBody:
        description: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInputModel'
        required: true
      responses:
        201:
          description: Search executed successfully. If results were found they are
            returned as a list. If no results are found then an empty list is returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseObject'
              example:
                message: "user created: 5d30cc07ef01bfcd6a089338, name: John Doe"
                response:
                  _id:  "5d30cc07ef01bfcd6a089338"
                  name: "John Doe"
                  age: 30
                  matches: []
                  interests: []
                  createdAt: "2019-07-18T19:44:07.672Z"
                  updatedAt: "2019-07-18T19:44:07.672Z"
                  __v: 0
        422:
          description: Invalide body data, either incorrectly formatted or missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
              example:
                message: "invalid body data supplied, name: null, age 30"
                error: {}
...
components:
  schemas:
    ResponseObject:
      type: object
      properties:
        message:
          type: string
        response:
          # oneOf
          $ref: '#/components/schemas/UserModel'
    UserInputModel:
      type: object
      properties:
        name: 
          type: string
        age: 
          type: integer
      example:
        name: "John Smith"
        age: 30
    UserModel:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        age:
          type: integer
        interests:
          type: string
        matches:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        __v:
          type: integer
    ErrorObject:
      type: object
      properties:
        message:
          type: string
        response:
          type: string
...